Add skipped test for lazy apps:
This test checks for loading an App while giving it a different
appname. The "appname" is the only attribute which can't be changed
because it controls how an App is created. Apps are not lazy, so
this won't work.

This is an idea that I hope we will be able to accomplish in the
future.